Also allow catalog.xml files in JARs, not only in SWCs.
Jangaroo uses the same catalog.xml format as Flash/Flex SWC, but packages
its artifacts as JARs.
It should not do harm to look for catalog.xml entries in the root directory of
JAR archives that are in the dependency list of a Flex module, too.
(cherry picked from commit 1176f3d)